requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-ollama>=0.0.3
langchain-chroma>=0.0.1
sentence-transformers>=2.2.2
einops>=0.8.0  # Required by nomic-ai/nomic-embed-text-v1
chromadb>=0.4.22
python-dotenv>=1.0.0
unstructured>=0.10.30
pdf2image>=1.16.3
python-magic>=0.4.27
tiktoken>=0.5.2
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
python-multipart>=0.0.6
pdfminer.six>=20221105
httpx==0.24.1

# Authentication dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Frontend dependencies
jinja2>=3.0.0
aiofiles>=0.8.0

# Testing dependencies
pytest
pytest-asyncio
pytest-httpx
requests